# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/tyokyo320/rateflow
  description: |-
    Exchange rate tracking service for UnionPay CNY/JPY rates

    This API provides access to historical and current exchange rates
    between Chinese Yuan (CNY) and Japanese Yen (JPY) from UnionPay.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: RateFlow API
  version: 1.0.0
paths:
  /api/rates:
    get:
      consumes:
      - application/json
      description: Retrieves the exchange rate for a given currency pair on a specific
        date
      parameters:
      - description: Currency pair (e.g., CNY/JPY, CNYJPY, or CNY-JPY)
        in: query
        name: pair
        required: true
        type: string
      - description: Date in YYYY-MM-DD format (e.g., 2025-01-15)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with rate data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get exchange rate for a specific date
      tags:
      - rates
  /api/rates/latest:
    get:
      consumes:
      - application/json
      description: Retrieves the most recent exchange rate for a given currency pair
      parameters:
      - description: Currency pair (e.g., CNY/JPY, CNYJPY, or CNY-JPY)
        in: query
        name: pair
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with rate data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get latest exchange rate
      tags:
      - rates
  /api/rates/list:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of exchange rates, optionally filtered
        by currency pair
      parameters:
      - description: Currency pair filter (e.g., CNY/JPY, CNYJPY, or CNY-JPY)
        in: query
        name: pair
        type: string
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: 'Items per page (default: 20, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response with paginated rate list and metadata
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List exchange rates with pagination
      tags:
      - rates
  /api/v1/rates:
    get:
      consumes:
      - application/json
      description: Retrieves the exchange rate for a given currency pair on a specific
        date
      parameters:
      - description: Currency pair (e.g., CNY/JPY, CNYJPY, or CNY-JPY)
        in: query
        name: pair
        required: true
        type: string
      - description: Date in YYYY-MM-DD format (e.g., 2025-01-15)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with rate data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get exchange rate for a specific date
      tags:
      - rates
  /api/v1/rates/latest:
    get:
      consumes:
      - application/json
      description: Retrieves the most recent exchange rate for a given currency pair
      parameters:
      - description: Currency pair (e.g., CNY/JPY, CNYJPY, or CNY-JPY)
        in: query
        name: pair
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success response with rate data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Rate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get latest exchange rate
      tags:
      - rates
  /api/v1/rates/list:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of exchange rates, optionally filtered
        by currency pair
      parameters:
      - description: Currency pair filter (e.g., CNY/JPY, CNYJPY, or CNY-JPY)
        in: query
        name: pair
        type: string
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: 'Items per page (default: 20, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response with paginated rate list and metadata
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: List exchange rates with pagination
      tags:
      - rates
  /health:
    get:
      description: Returns the health status of the API service
      produces:
      - application/json
      responses:
        "200":
          description: Health status response
          schema:
            additionalProperties: true
            type: object
      summary: Health check endpoint
      tags:
      - health
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Exchange rate operations
  name: rates
- description: Health check operations
  name: health